Rename point request schema and update imports
